# Conflicts:
#	src/main/res/values/strings.xml

+++ src/main/java/com/nextcloud/client/media/PlayerService.kt
@@ -139,7 +139,7 @@ class PlayerService : Service() {
     private fun onActionPlay(intent: Intent) {
         val user: User = intent.getParcelableExtra(EXTRA_USER) as User
         val file: OCFile = intent.getParcelableExtra(EXTRA_FILE) as OCFile
-        val startPos = intent.getIntExtra(EXTRA_START_POSITION_MS, 0)
+        val startPos = intent.getLongExtra(EXTRA_START_POSITION_MS, 0)
         val autoPlay = intent.getBooleanExtra(EXTRA_AUTO_PLAY, true)
         val item = PlaylistItem(file = file, startPositionMs = startPos, autoPlay = autoPlay, user = user)
         player.play(item)
